+++ domain/src/main/java/com/catshome/ClassJournal/domain/Group/GroupRepository.kt
@@ -8,7 +8,7 @@ interface GroupRepository {
  fun updateGroup(group: Group):Boolean
  fun deleteGroup(group: Group):Boolean
 
- fun getGroupById(uid:Int):Group
+ fun getGroupById(uid: Long):Group
  fun getGroupByName(name:String):Group
  fun getGroups(isDelete:Boolean):Flow<List<Group>>
 }
